backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.3
pydantic-settings==2.1.0
aiofiles==23.2.1
sqlmodel==0.0.22
numpy==1.26.4
scipy==1.11.4
opencv-python-headless==4.10.0.84
Pillow==10.4.0
typer==0.12.5
setuptools>=65.0.0
face-recognition==1.3.0
face-recognition-models @ git+https://github.com/ageitgey/face_recognition_models
pytest==8.3.2
httpx==0.27.2
